requirements.txt
addict==2.4.0
apex==0.9.10dev
basicsr==1.4.2
cityscapesscripts==2.2.2
decord==0.6.0
diffusers==0.16.0
einops==0.7.0
fairscale==0.4.4
ftfy==6.1.3
gradio==4.12.0
horovod==0.28.1
iopath==0.1.10
lmdb==1.4.1
matplotlib==3.7.4
mc==0.0
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.15.0
openai==1.6.1
opencv_python_headless==4.5.5.64
opendatasets==0.1.22
packaging==23.2
pandas==2.0.3
parrots==0.1.7
peft==0.7.1
Pillow==10.1.0
Pillow==10.1.0
plotly==5.18.0
prettytable==3.9.0
pycocoevalcap==1.2
pycocotools==2.0.7
pytest==7.4.4
python_magic==0.4.27
PyYAML==6.0.1
PyYAML==6.0.1
regex==2023.12.25
Requests==2.31.0
retrying==1.3.4
SceneGraphParser==0.1.0
scipy==1.11.4
setuptools==68.2.2
skimage==0.0
spacy==3.7.2
streamlit==1.29.0
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
tensorflow_hub==0.15.0
tensorflow_text==2.15.0
tifffile==2023.7.10
timm==0.4.12
torch==1.13.1+cu116
torch_tb_profiler==0.4.3
torchvision==0.14.1+cu116
tqdm==4.66.1
transformers==4.31.0
turbojpeg==0.0.2
visual_genome==1.1.1
webdataset==0.2.86
yapf==0.40.2
